+++ constants/definitions.py
@@ -2,4 +2,4 @@ import os
 
 CURR_DIR = os.path.dirname(__file__)
 ROOT_DIR = CURR_DIR[0:(len(CURR_DIR) - 9)]
-CSV_PATH = ROOT_DIR + '/csv'
+CSV_PATH = ROOT_DIR + '/csv_files'
